requirements.txt
streamlit
pandas
plotly
numpy
python-dotenv
schedule
supabase
selenium
webdriver-manager
requests
beautifulsoup4
